Inquiry view has been completed

+++ Shared/Sanaap.Shared/Dto/EvlRequestDto.cs
@@ -48,6 +48,6 @@ namespace Sanaap.Dto
 
         public string LostPlateNumber { get; set; }
 
-        public long Code { get; set; }
+        public int Code { get; set; }
     }
 }
